RoomNav: ledge-mantle synthetic flag (perf, Step 2 of 3)
User 2026-05-12 PM perf-optimization plan §Step 2. The
ComputePursuitSubgoal LedgeMantle path previously scanned
navData->ledgeAnchors every tick (O(N), up to ~600 anchors per
Kokiri-class room) to detect when the current subgoal corresponded
to a LedgeAnchor.topPos and the follower was at the matching
approachPos. The conceptual gap: BFS already knew the path
traversed a ledge edge (added in c680ad6f0) but waypoints carried
no metadata about WHICH edge type was used.

Fix: synthetic flag NODE_REACHED_VIA_LEDGE_GRAB (0x00010000) ORed
into waypointFlags by FindBestReachableSubgoalPath when the
reconstructed path traverses an approach→top ledge edge. Mirrors
the existing NODE_DROP_FROM_ABOVE pattern.

Widens NavPath::waypointFlags from uint16_t to uint32_t — the
lower 16 bits remain natural NodeFlags from NavNode::flags;
upper 16 bits carry synthetic flags. waypointFlags is runtime-
only (not persisted), so widening costs ~2 bytes per waypoint
in RAM and zero on disk. Updated all consumers (Follower.cpp,
FollowerRecorder.cpp, ActorTrail.cpp); ShouldZeroStickForEdge
keeps its uint16_t API and the caller passes a narrowed value
(safe — the function only inspects low-16-bit semantics).

AdjacencyCacheEntry gains a ledgeEdges set parallel to dropEdges;
BuildAdjacencyList collects approach→top pairs into it when
useLedgeAnchors=true (v1: always). Cache wrapper plumbs the set
through; path extraction in FindBestReachableSubgoalPath checks
it per-edge during the parents-array reverse-walk.

ComputePursuitSubgoal's ledge mantle is now a single flag check
on CurrentSubgoalFlags — O(1) per tick.

+++ soh/soh/Network/Anchor/AIFollower/FollowerRecorder.cpp
@@ -375,10 +375,10 @@ void CaptureFrame(const FollowerFrameContext& ctx) {
         anchor->followerNavPath.computedClimbMask);
     if (pathPresent) {
         size_t climbCount = 0;
-        uint16_t flagsAggregate = 0;  // OR of all remaining waypoint flags
+        uint32_t flagsAggregate = 0;  // OR of all remaining waypoint flags
         for (size_t i = anchor->followerNavPath.cursorIdx;
              i < anchor->followerNavPath.waypointFlags.size(); i++) {
-            const uint16_t f = anchor->followerNavPath.waypointFlags[i];
+            const uint32_t f = anchor->followerNavPath.waypointFlags[i];
             flagsAggregate |= f;
             if (f & ::AnchorNavRoom::NODE_CLIMB_ANY) {
                 climbCount++;
@@ -400,8 +400,8 @@ void CaptureFrame(const FollowerFrameContext& ctx) {
         nlohmann::json wpts = nlohmann::json::array();
         for (size_t i = 0; i < anchor->followerNavPath.waypoints.size(); i++) {
             const Vec3f& p = anchor->followerNavPath.waypoints[i];
-            uint16_t f = (i < anchor->followerNavPath.waypointFlags.size())
-                ? anchor->followerNavPath.waypointFlags[i] : (uint16_t)0;
+            uint32_t f = (i < anchor->followerNavPath.waypointFlags.size())
+                ? anchor->followerNavPath.waypointFlags[i] : (uint32_t)0;
             wpts.push_back({
                 {"i", static_cast<uint64_t>(i)},
                 {"x", p.x}, {"y", p.y}, {"z", p.z},
